refactor: extract screen components
