+++ src/App.js
@@ -28,6 +28,14 @@ import {
 // --- CORRECTION DE L'IMPORT POUR LE PLUGIN SOCIAL LOGIN ---
 import { SocialLogin } from '@capgo/capacitor-social-login';
 import { storage } from './storage';
+import DashboardScreen from './screens/DashboardScreen';
+import DiscoverScreen from './screens/DiscoverScreen';
+import FavoritesListScreen from './screens/FavoritesListScreen';
+import CompareScreen from './screens/CompareScreen';
+import ProfileManagementScreen from './screens/ProfileManagementScreen';
+import FriendsScreen from './screens/FriendsScreen';
+import NameDetailScreen from './screens/NameDetailScreen';
+import AuthScreen from './screens/AuthScreen';
 
 
 // --- CONFIGURATION FIREBASE (REMPLIR AVEC VOS INFORMATIONS) ---
@@ -43,8 +51,8 @@ const firebaseConfig = {
 
 // --- INITIALISATION DE FIREBASE ---
 const app = initializeApp(firebaseConfig);
-const auth = getAuth(app);
-const db = getFirestore(app);
+export const auth = getAuth(app);
+export const db = getFirestore(app);
 
 
 // --- ICONS (Identiques à la version précédente) ---
@@ -418,114 +426,6 @@ const AppLayout = () => {
 };
 
 // --- AUTHENTICATION SCREEN (MODIFIED) ---
-const AuthScreen = () => {
-    const [isLogin, setIsLogin] = useState(true);
-    const [email, setEmail] = useState('');
-    const [password, setPassword] = useState('');
-    const [error, setError] = useState('');
-    const [loading, setLoading] = useState(false);
-
-    const handleEmailAuth = async (e) => {
-        e.preventDefault();
-        setError('');
-        setLoading(true);
-        try {
-            if (isLogin) {
-                await signInWithEmailAndPassword(auth, email, password);
-            } else {
-                await createUserWithEmailAndPassword(auth, email, password);
-            }
-        } catch (err) {
-            setError(getFriendlyAuthError(err.code));
-        } finally {
-            setLoading(false);
-        }
-    };
-
-    const handleSocialLogin = async (provider) => {
-        setError('');
-        try {
-            // 1. Appeler le plugin natif
-            const result = await SocialLogin.signIn({ provider });
-
-            let credential;
-            // 2. Créer la bonne "credential" Firebase en fonction du fournisseur
-            if (provider === 'google') {
-                credential = GoogleAuthProvider.credential(result.token);
-            } else if (provider === 'facebook') {
-                credential = FacebookAuthProvider.credential(result.token);
-            } else {
-                throw new Error("Fournisseur non supporté");
-            }
-
-            // 3. Se connecter à Firebase
-            await signInWithCredential(auth, credential);
-
-        } catch (err) {
-            console.error(err);
-            setError(`La connexion avec ${provider} a échoué.`);
-        }
-    };
-
-    const getFriendlyAuthError = (code) => {
-        switch (code) {
-            case 'auth/user-not-found':
-            case 'auth/wrong-password':
-                return 'Email ou mot de passe incorrect.';
-            case 'auth/email-already-in-use':
-                return 'Cette adresse email est déjà utilisée.';
-            case 'auth/weak-password':
-                return 'Le mot de passe doit contenir au moins 6 caractères.';
-            case 'auth/invalid-email':
-                return 'Adresse email invalide.';
-            default:
-                return 'Une erreur est survenue. Veuillez réessayer.';
-        }
-    };
-
-    return (
-        <div className="bg-background min-h-screen flex flex-col items-center justify-center p-4 font-nunito">
-            <BackgroundDecorations />
-            <div className="w-full max-w-sm text-center">
-                <h1 className="text-4xl font-bold text-lavande-poudre mb-2">BabyNames</h1>
-                <p className="text-text-secondary mb-8">Trouvez le prénom parfait, ensemble.</p>
-
-                <Card className="text-left">
-                    <h2 className="text-2xl font-bold text-text-primary mb-6 text-center">{isLogin ? 'Connexion' : 'Inscription'}</h2>
-                    {error && <p className="bg-rose-saumon/20 text-rose-saumon-dark p-3 rounded-lg mb-4 text-center text-sm">{error}</p>}
-                    <form onSubmit={handleEmailAuth} className="space-y-4">
-                        <input type="email" value={email} onChange={e => setEmail(e.target.value)} placeholder="Adresse email" required className="w-full px-4 py-3 border border-subtle-border bg-background text-text-primary rounded-lg focus:outline-none focus:ring-2 focus:ring-lavande-poudre" />
-                        <input type="password" value={password} onChange={e => setPassword(e.target.value)} placeholder="Mot de passe" required className="w-full px-4 py-3 border border-subtle-border bg-background text-text-primary rounded-lg focus:outline-none focus:ring-2 focus:ring-lavande-poudre" />
-                        <Button type="submit" disabled={loading} className="w-full">
-                            {loading ? <LoadingSpinner small /> : (isLogin ? 'Se connecter' : "S'inscrire")}
-                        </Button>
-                    </form>
-                    <div className="relative my-6">
-                        <div className="absolute inset-0 flex items-center"><div className="w-full border-t border-subtle-border"></div></div>
-                        <div className="relative flex justify-center text-sm"><span className="px-2 bg-card-bg text-text-secondary">OU</span></div>
-                    </div>
-                    <div className="space-y-3">
-                        <button onClick={() => handleSocialLogin('google')} className="w-full flex items-center justify-center gap-3 py-3 px-4 border border-subtle-border rounded-lg hover:bg-hover-bg transition-colors">
-                            <img src="https://www.google.com/favicon.ico" alt="Google icon" className="h-5 w-5" />
-                            <span className="font-semibold text-text-primary">Continuer avec Google</span>
-                        </button>
-                        <button onClick={() => handleSocialLogin('facebook')} className="w-full flex items-center justify-center gap-3 py-3 px-4 border border-subtle-border rounded-lg hover:bg-hover-bg transition-colors">
-                            <svg className="h-5 w-5 text-[#1877F2]" fill="currentColor" viewBox="0 0 24 24"><path d="M22 12c0-5.52-4.48-10-10-10S2 6.48 2 12c0 4.84 3.44 8.87 8 9.8V15H8v-3h2V9.5C10 7.57 11.57 6 13.5 6H16v3h-2c-.55 0-1 .45-1 1v2h3l-.5 3h-2.5v6.8c4.56-.93 8-4.96 8-9.8z"></path></svg>
-                            <span className="font-semibold text-text-primary">Continuer avec Facebook</span>
-                        </button>
-                    </div>
-                    <p className="text-center mt-6 text-sm">
-                        <span className="text-text-secondary">{isLogin ? "Vous n'avez pas de compte ?" : 'Vous avez déjà un compte ?'}</span>
-                        <button onClick={() => { setIsLogin(!isLogin); setError('') }} className="font-semibold text-lavande-poudre hover:underline ml-1">
-                            {isLogin ? "S'inscrire" : 'Se connecter'}
-                        </button>
-                    </p>
-                </Card>
-            </div>
-        </div>
-    );
-};
-
 
 // --- UI COMPONENTS ---
 const LoadingSpinner = ({ small }) => <div className={`border-solid rounded-full animate-spin border-slate-200 dark:border-slate-700 ${small ? 'w-6 h-6 border-2 border-t-white' : 'w-12 h-12 border-4 border-t-lavande-poudre'}`}></div>;
@@ -601,431 +501,8 @@ const AnimatedBackground = ({ theme }) => {
 };
 
 // --- Reste des composants (inchangés) ---
-const DashboardScreen = ({ profile, allNames }) => { const liked = useMemo(() => profile.likedNames.map(id => allNames.find(n => n.id === id)).filter(Boolean), [profile.likedNames, allNames]); const m = useMemo(() => liked.filter(n => n.genre === 'M').length, [liked]); const boyPercent = liked.length > 0 ? Math.round((m / liked.length) * 100) : 0; const medalColors = ['#FFD700', '#C4C4C4', '#D9A77A']; return (<AnimatedPage className="space-y-6"><div className="text-center"><h1 className="text-3xl font-bold text-text-primary">Bonjour, {profile.displayName} !</h1><p className="text-text-secondary mt-1">Voici le résumé de vos découvertes.</p></div><Card><h2 className="font-bold text-lg mb-4 text-center text-text-primary">Statistiques</h2><div className="flex justify-around text-center"><div className="flex flex-col items-center space-y-1"><EyeIcon className="h-8 w-8 text-lavande-poudre" /><p className="text-3xl font-bold text-text-primary">{profile.seenNames.length}</p><p className="text-text-secondary text-sm">Prénoms vus</p></div><div className="flex flex-col items-center space-y-1"><HeartIcon className="h-8 w-8 text-rose-saumon" /><p className="text-3xl font-bold text-text-primary">{profile.likedNames.length}</p><p className="text-text-secondary text-sm">Coups de cœur</p></div></div></Card><Card><h2 className="font-bold text-lg mb-3 text-center text-text-primary">Répartition Garçon/Fille</h2><div className="w-full bg-rose-saumon-bg rounded-full h-4 relative overflow-hidden"><div className="bg-bleu-layette-bg h-4 rounded-l-full" style={{ width: `${boyPercent}%` }}></div></div><div className="flex justify-between text-xs text-text-secondary mt-1"><span>Garçon {boyPercent}%</span><span>Fille {100 - boyPercent}%</span></div></Card><Card><h2 className="font-bold text-lg mb-3 text-center text-text-primary">Nos petits trésors 🏆</h2>{profile.likedNames.length > 0 ? (<ul className="space-y-2">{profile.likedNames.slice(0, 3).map((nameId, index) => { const name = allNames.find(n => n.id === nameId); if (!name) return null; const genderBg = name.genre === 'M' ? 'bg-bleu-layette-bg' : 'bg-rose-saumon-bg'; return (<li key={name.id} className={`flex items-center space-x-4 p-3 rounded-xl ${genderBg} border-l-4`} style={{ borderColor: medalColors[index] }}><TrophyIcon color={medalColors[index]} className="h-7 w-7 flex-shrink-0" /><p className="font-semibold text-text-primary"><span className="text-lg">{name.prénom}</span>{profile.lastName && <span className="text-base ml-2 text-text-secondary">{profile.lastName}</span>}</p></li>) })}</ul>) : (<p className="text-center text-text-secondary py-4">Aimez des prénoms pour voir vos trésors !</p>)}</Card></AnimatedPage>); };
-const DiscoverScreen = ({ profile, allNames, updateProfile, origins }) => {
-    const [cardStack, setCardStack] = useState([]);
-    const [currentIndex, setCurrentIndex] = useState(0);
-    const [isFilterOpen, setFilterOpen] = useState(false);
-    const [filter, setFilter] = useState({ gender: 'Tous', origins: [], length: { type: 'Tous', min: 2, max: 15 } });
 
-    const memoizedCardStack = useMemo(() => {
-        if (allNames.length === 0 || !profile) return [];
-        const potentialNames = allNames.filter(n => {
-            // --- CORRECTION DE L'ERREUR ---
-            const genderMatch = (filter.gender === 'Tous' || n.genre === filter.gender);
-            const originMatch = (filter.origins.length === 0 || (n.filtre_global && filter.origins.includes(n.filtre_global)));
-            const nameLength = n.prénom.length;
-            const { type, min, max } = filter.length;
-            let lengthMatch = true;
-            switch (type) {
-                case 'Court': lengthMatch = nameLength >= 3 && nameLength <= 4; break;
-                case 'Moyen': lengthMatch = nameLength >= 5 && nameLength <= 6; break;
-                case 'Long': lengthMatch = nameLength >= 7; break;
-                case 'Personnalisé': lengthMatch = nameLength >= min && nameLength <= max; break;
-                default: lengthMatch = true;
-            }
-            return !profile.seenNames.includes(n.id) && genderMatch && originMatch && lengthMatch;
-        });
-        return [...potentialNames].sort(() => Math.random() - 0.5);
-    }, [filter, allNames, profile.id]);
 
-    useEffect(() => { setCardStack(memoizedCardStack); setCurrentIndex(0); }, [memoizedCardStack]);
-
-    const swiped = (direction, name) => {
-        const wasLikedBeforeSwipe = profile.likedNames.includes(name.id);
-        const newHistory = [...profile.history, { nameId: name.id, wasLiked: wasLikedBeforeSwipe }];
-        const isLike = direction === 'right';
-        const newLikedNames = isLike ? Array.from(new Set([...profile.likedNames, name.id])) : profile.likedNames.filter(id => id !== name.id);
-        updateProfile({ history: newHistory, likedNames: newLikedNames, seenNames: Array.from(new Set([...profile.seenNames, name.id])) });
-        setCurrentIndex(prev => prev + 1);
-    };
-
-    const handleUndo = useCallback(() => {
-        if (currentIndex === 0 || profile.history.length === 0) return;
-        const lastAction = profile.history[profile.history.length - 1];
-        if (!lastAction) return;
-        const cardToRestore = cardStack[currentIndex - 1];
-        if (!cardToRestore || cardToRestore.id !== lastAction.nameId) { console.error("Undo error"); return; }
-        const newHistory = profile.history.slice(0, -1);
-        const newSeenNames = profile.seenNames.filter(id => id !== lastAction.nameId);
-        let newLikedNames;
-        if (lastAction.wasLiked) {
-            newLikedNames = Array.from(new Set([...profile.likedNames, lastAction.nameId]));
-        } else {
-            newLikedNames = profile.likedNames.filter(id => id !== lastAction.nameId);
-        }
-        updateProfile({ history: newHistory, seenNames: newSeenNames, likedNames: newLikedNames });
-        setCurrentIndex(prev => prev - 1);
-    }, [currentIndex, profile, cardStack, updateProfile]);
-
-    if (isFilterOpen) {
-        return <FilterScreen filter={filter} origins={origins} onSetFilter={setFilter} onLaunch={() => setFilterOpen(false)} allNames={allNames} />;
-    }
-
-    const visibleStack = cardStack.slice(currentIndex, currentIndex + 4);
-
-    return (
-        <AnimatedPage className="h-full flex flex-col">
-            <div className="flex justify-end items-center space-x-3 mb-4">
-                <button
-                    onClick={handleUndo}
-                    aria-label="Annuler le dernier choix"
-                    className="p-3 bg-card-bg text-text-primary rounded-full shadow-soft disabled:opacity-50 disabled:cursor-not-allowed transition-all hover:bg-hover-bg"
-                    disabled={currentIndex === 0 || profile.history.length === 0}
-                >
-                    <UndoIcon />
-                </button>
-                <button
-                    onClick={() => setFilterOpen(true)}
-                    aria-label="Ouvrir les filtres"
-                    className="p-3 bg-card-bg text-text-primary rounded-full shadow-soft transition-all hover:bg-hover-bg"
-                >
-                    <FilterIcon />
-                </button>
-            </div>
-            <div className="flex-grow relative flex items-center justify-center -mt-8 overflow-hidden">
-                {visibleStack.length > 0 ? (
-                    visibleStack.reverse().map((name, index) => {
-                        const stackIndex = visibleStack.length - 1 - index;
-                        const isTop = stackIndex === 0;
-                        const style = { zIndex: 100 - stackIndex, transform: `scale(${1 - stackIndex * 0.05}) translateY(-${stackIndex * 10}px)`, opacity: stackIndex > 2 ? 0 : 1, transition: 'transform 0.4s cubic-bezier(0.175, 0.885, 0.32, 1.275), opacity 0.4s ease-out', };
-                        return (<SwipeableCard key={name.id} name={name} lastName={profile.lastName} onSwipe={(dir) => swiped(dir, name)} isTop={isTop} style={style} />);
-                    })
-                ) : (
-                    <Card className="text-center">
-                        <h3 className="text-xl font-semibold text-text-primary">C'est tout pour le moment !</h3>
-                        <p className="text-text-secondary mt-2">Essayez de modifier vos filtres pour découvrir de nouveaux prénoms.</p>
-                    </Card>
-                )}
-            </div>
-        </AnimatedPage>
-    );
-};
-const FavoritesListScreen = ({ profile, allNames, updateProfile, setDetailedName }) => { const [favoriteItems, setFavoriteItems] = useState([]); const [draggedItem, setDraggedItem] = useState(null); useEffect(() => { setFavoriteItems([...profile.likedNames].reverse()); }, [profile.likedNames]); const handleRemove = (nameId) => { const newLikedNames = profile.likedNames.filter(id => id !== nameId); updateProfile({ likedNames: newLikedNames }); }; const handleDragStart = (e, index) => { setDraggedItem(index); e.dataTransfer.effectAllowed = 'move'; }; const handleDragOver = (e, index) => { e.preventDefault(); }; const handleDrop = (e, dropIndex) => { const draggedId = favoriteItems[draggedItem]; const newItems = [...favoriteItems]; newItems.splice(draggedItem, 1); newItems.splice(dropIndex, 0, draggedId); setFavoriteItems(newItems); updateProfile({ likedNames: [...newItems].reverse() }); }; const handleDragEnd = () => { setDraggedItem(null); }; return (<AnimatedPage><h1 className="text-3xl font-bold text-text-primary mb-2">Vos Trésors</h1><p className="text-text-secondary mb-6 -mt-1">Organisez votre liste en glissant-déposant les prénoms.</p>{favoriteItems.length > 0 ? (<ul className="space-y-3">{favoriteItems.map((id, index) => { const name = allNames.find(n => n.id === id); if (!name) return null; return (<FavoriteItem key={id} name={name} lastName={profile.lastName} index={index} isBeingDragged={draggedItem === index} onRemove={handleRemove} onViewDetails={setDetailedName} onDragStart={(e) => handleDragStart(e, index)} onDragOver={(e) => handleDragOver(e, index)} onDrop={(e) => handleDrop(e, index)} onDragEnd={handleDragEnd} />); })}</ul>) : (<Card className="text-center"><HeartIcon className="h-16 w-16 mx-auto text-slate-300 dark:text-slate-600" /><h3 className="text-xl font-semibold text-text-primary mt-4">Aucun trésor pour l'instant</h3><p className="text-text-secondary mt-2">Allez dans "Trouvailles" pour découvrir des prénoms !</p></Card>)}</AnimatedPage>); };
-const CompareScreen = ({ profile, allNames, updateProfile, onNavigate, friends, isFriendsLoading }) => {
-    const [selectedFriend, setSelectedFriend] = useState(null);
-    const [friendData, setFriendData] = useState(null);
-    const [loading, setLoading] = useState(false);
-
-    const handleSelectFriend = async (friend) => {
-        setLoading(true);
-        setSelectedFriend(friend);
-        const userRef = doc(db, "users", friend.uid);
-        const docSnap = await getDoc(userRef);
-        if (docSnap.exists()) {
-            const data = docSnap.data();
-            setFriendData({
-                ...data,
-                likedNames: data.likedNames || [],
-                lastName: data.lastName || '',
-            });
-        } else {
-            console.error("Friend data not found!");
-            setFriendData(null);
-        }
-        setLoading(false);
-    };
-
-    const handleToggleLike = (nameId) => {
-        const currentLiked = new Set(profile.likedNames);
-        if (currentLiked.has(nameId)) {
-            currentLiked.delete(nameId);
-        } else {
-            currentLiked.add(nameId);
-        }
-        updateProfile({ likedNames: Array.from(currentLiked) });
-    };
-
-    if (isFriendsLoading) {
-        return <div className="flex justify-center items-center h-full"><LoadingSpinner /></div>;
-    }
-
-    if (!selectedFriend || !friendData) {
-        return (
-            <AnimatedPage className="space-y-4">
-                <h1 className="text-3xl font-bold text-center text-text-primary">Comparer avec...</h1>
-                {friends.length > 0 ? friends.map(f => (
-                    <Card key={f.uid} onClick={() => handleSelectFriend(f)} className="flex items-center justify-between p-4 transition-all hover:bg-hover-bg cursor-pointer">
-                        <div className="flex items-center space-x-4">
-                            <img src={f.photoURL || `https://i.pravatar.cc/150?u=${f.uid}`} alt={f.displayName} className="h-12 w-12 rounded-full" />
-                            <p className="font-bold text-text-primary">{f.displayName}</p>
-                        </div>
-                        <span className="text-lavande-poudre font-bold">&rarr;</span>
-                    </Card>
-                )) : (
-                    <Card className="text-center">
-                        <p className="text-text-secondary">Vous n'avez pas encore d'amis.</p>
-                        <Button onClick={() => onNavigate('friends')} className="mt-4">Ajouter des amis</Button>
-                    </Card>
-                )}
-            </AnimatedPage>
-        );
-    }
-
-    if (loading) {
-        return <div className="flex justify-center items-center h-full"><LoadingSpinner /></div>;
-    }
-
-    const commonLikedNames = profile.likedNames.filter(id => friendData.likedNames.includes(id));
-    const myUniqueLiked = profile.likedNames.filter(id => !friendData.likedNames.includes(id));
-    const friendUniqueLiked = friendData.likedNames.filter(id => !profile.likedNames.includes(id));
-    const getNameById = (id) => allNames.find(n => n.id === id);
-
-    return (
-        <AnimatedPage className="space-y-6">
-            <button onClick={() => { setSelectedFriend(null); setFriendData(null); }} className="text-lavande-poudre font-semibold flex items-center space-x-2 group mb-4">
-                <span className="transform transition-transform group-hover:-translate-x-1">&larr;</span>
-                <span>Retour</span>
-            </button>
-            <div className="text-center">
-                <h1 className="text-3xl font-bold text-text-primary">Accord Parfait</h1>
-                <p className="text-text-secondary mt-1">Vous et <span className="font-bold">{selectedFriend.displayName}</span></p>
-            </div>
-            <Card>
-                <h2 className="text-xl font-bold mb-4 text-text-primary text-center">💖 Matchs ({commonLikedNames.length})</h2>
-                {commonLikedNames.length > 0 ? (
-                    <ul className="space-y-2">
-                        {commonLikedNames.map(getNameById).filter(Boolean).map(name => <CompareRow key={name.id} name={name} lastName={profile.lastName} isLiked={true} onToggleLike={() => handleToggleLike(name.id)} />)}
-                    </ul>
-                ) : (
-                    <p className="text-text-secondary text-center">Aucun prénom en commun pour l'instant.</p>
-                )}
-            </Card>
-            <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
-                <Card>
-                    <h2 className="text-xl font-bold mb-4 text-text-primary">Vos trésors ({myUniqueLiked.length})</h2>
-                    {myUniqueLiked.length > 0 ? (
-                        <ul className="space-y-2">
-                            {myUniqueLiked.map(getNameById).filter(Boolean).map(name => <CompareRow key={name.id} name={name} lastName={profile.lastName} isLiked={true} onToggleLike={() => handleToggleLike(name.id)} />)}
-                        </ul>
-                    ) : (
-                        <p className="text-text-secondary">...</p>
-                    )}
-                </Card>
-                <Card>
-                    <h2 className="text-xl font-bold mb-4 text-text-primary">{friendData.displayName} ({friendUniqueLiked.length})</h2>
-                    {friendUniqueLiked.length > 0 ? (
-                        <ul className="space-y-2">
-                            {friendUniqueLiked.map(getNameById).filter(Boolean).map(name => <CompareRow key={name.id} name={name} lastName={friendData.lastName} isLiked={profile.likedNames.includes(name.id)} onToggleLike={() => handleToggleLike(name.id)} />)}
-                        </ul>
-                    ) : (
-                        <p className="text-text-secondary">...</p>
-                    )}
-                </Card>
-            </div>
-        </AnimatedPage>
-    );
-};
-
-const ProfileManagementScreen = ({ profile, updateProfile, theme, setTheme }) => {
-    const { user } = useAuth();
-    const [displayName, setDisplayName] = useState('');
-    const [lastName, setLastName] = useState('');
-
-    useEffect(() => {
-        if (profile) {
-            setDisplayName(profile.displayName);
-            setLastName(profile.lastName || '');
-        }
-    }, [profile]);
-
-    const handleProfileInfoSave = () => {
-        if (profile) updateProfile({ displayName, lastName });
-    };
-
-    const handleLogout = async () => {
-        try {
-            await signOut(auth);
-        } catch (error) {
-            console.error("Error signing out:", error);
-        }
-    };
-
-    return (
-        <AnimatedPage className="space-y-6">
-            <Card>
-                <h2 className="text-xl font-bold mb-4 text-text-primary">Mon Compte</h2>
-                <div className="flex items-center space-x-4 mb-6">
-                    <img src={user.photoURL || `https://i.pravatar.cc/150?u=${user.uid}`} alt="User" className="h-16 w-16 rounded-full" />
-                    <div>
-                        <p className="font-bold text-text-primary">{user.displayName}</p>
-                        <p className="text-sm text-text-secondary">{user.email}</p>
-                    </div>
-                </div>
-                <Button color="red" onClick={handleLogout} className="w-full mb-6 flex items-center justify-center gap-2"><LogoutIcon /> Se déconnecter</Button>
-            </Card>
-            <Card>
-                <h2 className="text-xl font-bold mb-4 text-text-primary">Informations</h2>
-                <div className="space-y-4">
-                    <div>
-                        <label htmlFor="display-name" className="block text-sm font-medium text-text-secondary mb-1">Nom d'utilisateur</label>
-                        <input id="display-name" type="text" value={displayName} onChange={e => setDisplayName(e.target.value)} className="w-full px-4 py-3 border border-subtle-border bg-hover-bg text-text-primary rounded-lg focus:outline-none focus:ring-2 focus:ring-lavande-poudre" />
-                    </div>
-                    <div>
-                        <label htmlFor="last-name" className="block text-sm font-medium text-text-secondary mb-1">Nom de famille (pour les matchs)</label>
-                        <input id="last-name" type="text" value={lastName} onChange={e => setLastName(e.target.value)} className="w-full px-4 py-3 border border-subtle-border bg-hover-bg text-text-primary rounded-lg focus:outline-none focus:ring-2 focus:ring-lavande-poudre" />
-                    </div>
-                    <Button onClick={handleProfileInfoSave} className="w-full">Enregistrer</Button>
-                </div>
-            </Card>
-            <Card>
-                <h2 className="text-xl font-bold mb-4 text-text-primary">Préférences</h2>
-                <ThemeToggle theme={theme} setTheme={setTheme} />
-            </Card>
-        </AnimatedPage>
-    );
-};
-
-const FriendsScreen = ({
-    friends,
-    requests,
-    sentRequests,
-    isFriendsLoading,
-    onNavigate,
-    setRequests,
-    setFriends,
-    currentUser,
-    setSentRequests
-}) => {
-    const [isSearchModalOpen, setSearchModalOpen] = useState(false);
-
-    const handleRequestAction = async (friendshipId, newStatus) => {
-        const request = requests.find(r => r.id === friendshipId);
-        if (!request || !request.user) return;
-
-        setRequests(prev => prev.filter(req => req.id !== friendshipId));
-        if (newStatus === 'accepted') {
-            setFriends(prev => [...prev, request.user]);
-        }
-
-        const friendshipRef = doc(db, 'friendships', friendshipId);
-        try {
-            if (newStatus === 'accepted') {
-                await updateDoc(friendshipRef, { status: 'accepted' });
-            } else {
-                await deleteDoc(friendshipRef);
-            }
-        } catch (error) {
-            console.error("Erreur lors de l'action sur la demande d'ami:", error);
-        }
-    };
-
-    const cancelSentRequest = async (friendshipId) => {
-        setSentRequests(prev => prev.filter(req => req.id !== friendshipId));
-        const friendshipRef = doc(db, 'friendships', friendshipId);
-        try {
-            await deleteDoc(friendshipRef);
-        } catch (error) {
-            console.error("Erreur lors de l'annulation de la demande:", error);
-        }
-    };
-
-    const addSentRequestOptimistically = (sentRequest) => {
-        setSentRequests(currentSentRequests => [sentRequest, ...currentSentRequests]);
-        setSearchModalOpen(false);
-    };
-
-    const removeFriend = async (friendId) => {
-        const confirmed = window.confirm("Êtes-vous sûr de vouloir supprimer cet ami ?");
-        if (!confirmed) return;
-        
-        setFriends(prev => prev.filter(f => f.uid !== friendId));
-        
-        try {
-            const userIds = [currentUser.uid, friendId].sort();
-            const friendshipId = userIds.join('_');
-            const friendshipRef = doc(db, 'friendships', friendshipId);
-            await deleteDoc(friendshipRef);
-        } catch (error) {
-            console.error("Erreur lors de la suppression de l'ami:", error);
-        }
-    };
-
-    if (isFriendsLoading) {
-        return <div className="flex justify-center items-center h-full"><LoadingSpinner /></div>;
-    }
-
-    return (
-        <AnimatedPage className="space-y-6">
-            <div className="flex justify-between items-center">
-                <h1 className="text-3xl font-bold text-text-primary">Mes Amis</h1>
-                <Button onClick={() => setSearchModalOpen(true)} className="flex items-center gap-2">
-                    <UserPlusIcon /> Ajouter
-                </Button>
-            </div>
-
-            {requests.length > 0 && (
-                <Card>
-                    <h2 className="text-xl font-bold mb-4 text-text-primary">Demandes d'amis</h2>
-                    <ul className="space-y-3">
-                        {requests.map(req => (
-                            <li key={req.id} className="flex items-center justify-between p-3 bg-hover-bg rounded-lg">
-                                <div className="flex items-center gap-3">
-                                    <img src={req.user?.photoURL || `https://i.pravatar.cc/150?u=${req.user?.uid}`} alt={req.user?.displayName} className="h-10 w-10 rounded-full" />
-                                    <span className="font-semibold text-text-primary">{req.user?.displayName}</span>
-                                </div>
-                                <div className="flex gap-2">
-                                    <Button onClick={() => handleRequestAction(req.id, 'accepted')} color="emerald" className="!py-2 !px-4">Accepter</Button>
-                                    <Button onClick={() => handleRequestAction(req.id, 'declined')} color="red" className="!py-2 !px-4">Refuser</Button>
-                                </div>
-                            </li>
-                        ))}
-                    </ul>
-                </Card>
-            )}
-
-            {sentRequests.length > 0 && (
-                <Card>
-                    <h2 className="text-xl font-bold mb-4 text-text-primary">Demandes envoyées</h2>
-                    <ul className="space-y-3">
-                        {sentRequests.map(req => (
-                            <li key={req.id} className="flex items-center justify-between p-3 bg-hover-bg rounded-lg">
-                                <div className="flex items-center gap-3">
-                                    <img src={req.user?.photoURL || `https://i.pravatar.cc/150?u=${req.user?.uid}`} alt={req.user?.displayName} className="h-10 w-10 rounded-full" />
-                                    <div>
-                                        <span className="font-semibold text-text-primary">{req.user?.displayName}</span>
-                                        <p className="text-sm text-text-secondary">En attente</p>
-                                    </div>
-                                </div>
-                                <Button onClick={() => cancelSentRequest(req.id)} color="red" className="!py-2 !px-4">Annuler</Button>
-                            </li>
-                        ))}
-                    </ul>
-                </Card>
-            )}
-
-            <Card>
-                <h2 className="text-xl font-bold mb-4 text-text-primary">Ma liste d'amis</h2>
-                {friends.length > 0 ? (
-                    <ul className="space-y-3">
-                        {friends.map(friend => (
-                            <li key={friend.uid} className="flex items-center justify-between p-3 bg-hover-bg rounded-lg">
-                                <div className="flex items-center gap-3">
-                                    <img src={friend.photoURL || `https://i.pravatar.cc/150?u=${friend.uid}`} alt={friend.displayName} className="h-10 w-10 rounded-full" />
-                                    <span className="font-semibold text-text-primary">{friend.displayName}</span>
-                                </div>
-                                <div className="flex items-center gap-2">
-                                    <Button onClick={() => onNavigate('compare')} className="!py-2 !px-4">Comparer</Button>
-                                    <button
-                                        onClick={() => removeFriend(friend.uid)}
-                                        aria-label="Supprimer l'ami"
-                                        className="text-rose-saumon p-2 rounded-full hover:bg-black/5 dark:hover:bg-white/5 transition-colors"
-                                    >
-                                        <TrashIcon className="h-5 w-5" />
-                                    </button>
-                                </div>
-                            </li>
-                        ))}
-                    </ul>
-                ) : (
-                    <p className="text-text-secondary text-center py-4">Vous n'avez pas encore d'amis. Ajoutez-en pour commencer à comparer !</p>
-                )}
-            </Card>
-
-            {isSearchModalOpen && <SearchFriendModal onClose={() => setSearchModalOpen(false)} currentUser={currentUser} friends={friends} onFriendRequestSent={addSentRequestOptimistically} />}
-        </AnimatedPage>
-    );
-};
 
 
 const SearchFriendModal = ({ onClose, currentUser, friends, onFriendRequestSent }) => {
@@ -1119,7 +596,6 @@ const SearchFriendModal = ({ onClose, currentUser, friends, onFriendRequestSent 
 
 
 
-const FilterScreen = ({ filter: current, origins, onSetFilter, onLaunch, allNames }) => { const handleOriginToggle = (origin) => { const newOrigins = [...current.origins]; const index = newOrigins.indexOf(origin); if (index > -1) { newOrigins.splice(index, 1); } else { newOrigins.push(origin); } onSetFilter({ ...current, origins: newOrigins }); }; const resetOrigins = () => { onSetFilter({ ...current, origins: [] }); }; const handleLengthTypeChange = (type) => { onSetFilter({ ...current, length: { ...current.length, type } }); }; const handleLengthRangeChange = ({ min, max }) => { onSetFilter({ ...current, length: { ...current.length, type: 'Personnalisé', min, max } }); }; const lengthTypes = [{ key: 'Tous', label: 'Toutes' }, { key: 'Court', label: 'Courts (3-4)' }, { key: 'Moyen', label: 'Moyens (5-6)' }, { key: 'Long', label: 'Longs (7+)' }, { key: 'Personnalisé', label: 'Personnalisé' },]; const filteredCount = useMemo(() => { if (!allNames) return 0; return allNames.filter(n => { const genderMatch = (current.gender === 'Tous' || n.genre === current.gender); const originMatch = (current.origins.length === 0 || (n.filtre_global && current.origins.includes(n.filtre_global))); const nameLength = n.prénom.length; const { type, min, max } = current.length; let lengthMatch = true; switch (type) { case 'Court': lengthMatch = nameLength >= 3 && nameLength <= 4; break; case 'Moyen': lengthMatch = nameLength >= 5 && nameLength <= 6; break; case 'Long': lengthMatch = nameLength >= 7; break; case 'Personnalisé': lengthMatch = nameLength >= min && nameLength <= max; break; default: lengthMatch = true; } return genderMatch && originMatch && lengthMatch; }).length; }, [allNames, current]); const totalCount = allNames.length; return (<> <AnimatedPage className="space-y-6 pb-24"> <Card> <h2 className="text-xl font-bold mb-4 text-text-primary">Genre</h2> <div className="flex justify-center space-x-2"> <Chip label="Garçon" selected={current.gender === 'M'} onClick={() => onSetFilter({ ...current, gender: 'M' })} /> <Chip label="Fille" selected={current.gender === 'F'} onClick={() => onSetFilter({ ...current, gender: 'F' })} /> <Chip label="Tous" selected={current.gender === 'Tous'} onClick={() => onSetFilter({ ...current, gender: 'Tous' })} /> </div> </Card> <Card> <div className="flex justify-between items-center mb-4"> <h2 className="text-xl font-bold text-text-primary">Origine(s) {current.origins.length > 0 ? `(${current.origins.length})` : ''}</h2> {current.origins.length > 0 && (<button onClick={resetOrigins} className="text-sm font-semibold text-lavande-poudre hover:underline">Tout désélectionner</button>)} </div> <div className="max-h-48 overflow-y-auto pr-2"> <div className="flex flex-wrap gap-2"> <Chip label="Toutes" selected={current.origins.length === 0} onClick={resetOrigins} /> {origins.map(origin => (<Chip key={origin} label={origin} selected={current.origins.includes(origin)} onClick={() => handleOriginToggle(origin)} />))} </div> </div> </Card> <Card> <h2 className="text-xl font-bold mb-4 text-text-primary">Longueur du prénom</h2> <div className="flex flex-wrap justify-center gap-2"> {lengthTypes.map(({ key, label }) => (<Chip key={key} label={label} selected={current.length.type === key} onClick={() => handleLengthTypeChange(key)} />))} </div> {current.length.type === 'Personnalisé' && (<div className="mt-6 animate-fade-in-up"> <RangeSlider min={2} max={15} currentMin={current.length.min} currentMax={current.length.max} onChange={handleLengthRangeChange} /> </div>)} </Card> <Card className="text-center mt-6 bg-card-bg/80 backdrop-blur-sm"> <p className="text-text-secondary text-sm sm:text-base"> <span className="font-bold text-text-primary">{filteredCount}</span> {filteredCount > 1 ? "prénoms correspondent" : "prénom correspond"} <br /> sur <span className="font-bold text-text-primary">{totalCount}</span> prénoms disponibles. </p> </Card> </AnimatedPage> <FAB icon={SearchIcon} onClick={onLaunch} ariaLabel="Appliquer les filtres" /> </>); };
 const SwipeableCard = ({ name, lastName, onSwipe, isTop, style }) => { const cardRef = useRef(); const nameRef = useRef(); const [feedback, setFeedback] = useState(null); const startPoint = useRef({ x: 0, y: 0 }); const isDragging = useRef(false); useEffect(() => { const resizeText = () => { const textEl = nameRef.current; const containerEl = textEl?.parentElement; if (!textEl || !containerEl) return; requestAnimationFrame(() => { const containerWidth = containerEl.offsetWidth; if (containerWidth <= 0) return; let fontSize = 64; textEl.style.fontSize = `${fontSize}px`; while (textEl.scrollWidth > containerWidth - 8 && fontSize > 16) { fontSize--; textEl.style.fontSize = `${fontSize}px`; } }); }; resizeText(); window.addEventListener('resize', resizeText); return () => { window.removeEventListener('resize', resizeText); }; }, [name.prénom, lastName]); const handleGestureStart = (e) => { if (!isTop) return; startPoint.current = { x: e.clientX ?? e.touches[0].clientX, y: e.clientY ?? e.touches[0].clientY }; isDragging.current = true; if (cardRef.current) { cardRef.current.style.transition = 'transform 0.1s'; } }; const handleGestureMove = (e) => { if (!isDragging.current || !isTop || !cardRef.current) return; const currentX = e.clientX ?? e.touches[0].clientX; const deltaX = currentX - startPoint.current.x; const rotation = deltaX / 20; cardRef.current.style.transform = `${style.transform} translateX(${deltaX}px) rotate(${rotation}deg)`; if (deltaX > 20) setFeedback('like'); else if (deltaX < -20) setFeedback('nope'); else setFeedback(null); }; const handleGestureEnd = (e) => { if (!isDragging.current || !isTop || !cardRef.current) return; isDragging.current = false; const deltaX = (e.clientX ?? e.changedTouches[0].clientX) - startPoint.current.x; if (Math.abs(deltaX) > 100) { const direction = deltaX > 0 ? 'right' : 'left'; const flyOutX = (deltaX > 0 ? 1 : -1) * (window.innerWidth); const rotation = deltaX / 15; cardRef.current.style.transition = 'transform 0.5s ease-out, opacity 0.5s ease-out'; cardRef.current.style.transform = `translateX(${flyOutX}px) rotate(${rotation * 2}deg)`; cardRef.current.style.opacity = '0'; setTimeout(() => onSwipe(direction), 500); } else { cardRef.current.style.transition = 'transform 0.4s cubic-bezier(0.175, 0.885, 0.32, 1.275)'; cardRef.current.style.transform = style.transform; setFeedback(null); } }; const triggerSwipe = (direction) => { if (!isTop || !cardRef.current) return; const flyOutX = (direction === 'right' ? 1 : -1) * window.innerWidth; const rotation = direction === 'right' ? 15 : -15; setFeedback(direction === 'right' ? 'like' : 'nope'); cardRef.current.style.transition = 'transform 0.6s cubic-bezier(0.68, -0.55, 0.27, 1.55), opacity 0.6s ease-out'; cardRef.current.style.transform = `translateX(${flyOutX}px) rotate(${rotation}deg)`; cardRef.current.style.opacity = '0'; setTimeout(() => onSwipe(direction), 600); }; const genderBg = name.genre === 'M' ? 'bg-bleu-layette-bg' : 'bg-rose-saumon-bg'; return (<div ref={cardRef} className={`absolute touch-none ${isTop ? 'cursor-grab active:cursor-grabbing pointer-events-auto' : 'pointer-events-none'}`} style={style} onMouseDown={handleGestureStart} onMouseMove={handleGestureMove} onMouseUp={handleGestureEnd} onMouseLeave={handleGestureEnd} onTouchStart={handleGestureStart} onTouchMove={handleGestureMove} onTouchEnd={handleGestureEnd}><Card className={`w-80 h-[450px] flex flex-col justify-between items-center relative overflow-hidden border border-subtle-border ${genderBg}`}><div className={`absolute inset-0 flex items-center justify-center transition-opacity duration-300 pointer-events-none ${feedback === 'like' ? 'opacity-100' : 'opacity-0'}`}><div className="border-8 border-vert-menthe rounded-lg transform -rotate-12"><h3 className="text-vert-menthe text-4xl font-black p-2 tracking-widest">OUI !</h3></div></div><div className={`absolute inset-0 flex items-center justify-center transition-opacity duration-300 pointer-events-none ${feedback === 'nope' ? 'opacity-100' : 'opacity-0'}`}><div className="border-8 border-rose-saumon rounded-lg transform rotate-12"><h3 className="text-rose-saumon text-4xl font-black p-2 tracking-widest">NON</h3></div></div><div className="text-center pt-8 px-4 overflow-hidden flex-grow flex flex-col w-full"><div className="w-full flex justify-center items-center h-28"><h2 ref={nameRef} className="font-bold text-text-primary leading-none whitespace-nowrap text-center">{name.prénom}{lastName && <span className="text-[0.8em] font-semibold ml-2 text-text-secondary">{lastName}</span>}</h2></div><div className="w-full flex-grow my-4 overflow-y-auto description-scroll pr-2"><p className="italic text-text-secondary text-base">"{name.signification}"</p></div></div><div className="w-full px-4 pb-4 flex-shrink-0"><div className="flex justify-center items-center mb-4"><div className="flex items-center space-x-2 text-text-secondary bg-card-bg/50 px-3 py-1 rounded-full border border-subtle-border">{name.genre === 'M' ? <MaleSymbolIcon className="h-4 w-4 text-bleu-layette" /> : <FemaleSymbolIcon className="h-4 w-4 text-rose-saumon" />}<p className="text-sm">Origine: {name.origines.join(', ')}</p></div></div><div className="flex justify-around w-full"><button onClick={() => triggerSwipe('left')} className="p-4 bg-card-bg rounded-full shadow-soft transform transition hover:scale-110 active:scale-95"><TrashIcon className="h-9 w-9 text-rose-saumon" /></button><button onClick={() => triggerSwipe('right')} className="p-4 bg-card-bg rounded-full shadow-soft transform transition hover:scale-110 active:scale-95"><HeartIcon className="h-9 w-9 text-vert-menthe" /></button></div></div></Card></div>); };
 const RangeSlider = ({ min, max, currentMin, currentMax, onChange }) => { const minValRef = useRef(currentMin); const maxValRef = useRef(currentMax); const range = useRef(null); const getPercent = useCallback((value) => Math.round(((value - min) / (max - min)) * 100), [min, max]); useEffect(() => { minValRef.current = currentMin; maxValRef.current = currentMax; }, [currentMin, currentMax]); useEffect(() => { const minPercent = getPercent(currentMin); const maxPercent = getPercent(currentMax); if (range.current) { range.current.style.left = `${minPercent}%`; range.current.style.width = `${maxPercent - minPercent}%`; } }, [currentMin, currentMax, getPercent]); return (<div className="w-full pt-4"> <div className="relative h-10 flex items-center justify-center"> <input type="range" min={min} max={max} value={currentMin} onChange={(event) => { const value = Math.min(Number(event.target.value), maxValRef.current - 1); onChange({ min: value, max: maxValRef.current }); }} className="thumb thumb--left" style={{ zIndex: currentMin > max - 100 ? '5' : '3' }} /> <input type="range" min={min} max={max} value={currentMax} onChange={(event) => { const value = Math.max(Number(event.target.value), minValRef.current + 1); onChange({ min: minValRef.current, max: value }); }} className="thumb thumb--right" /> <div className="relative w-full"> <div className="slider__track" /> <div ref={range} className="slider__range" /> <div className="slider__left-value">{currentMin}</div> <div className="slider__right-value">{currentMax}</div> </div> </div> <div className="flex justify-between text-xs text-text-secondary mt-2"> <span>{min} lettres</span> <span>{max} lettres</span> </div> </div>); };
 const FavoriteItem = ({ name, lastName, index, onRemove, onViewDetails, onDragStart, onDragOver, onDrop, onDragEnd, isBeingDragged }) => {
@@ -1166,46 +642,6 @@ const FavoriteItem = ({ name, lastName, index, onRemove, onViewDetails, onDragSt
     );
 };
 const CompareRow = ({ name, lastName, isLiked, onToggleLike }) => { const genderBg = name.genre === 'M' ? 'bg-bleu-layette-bg' : 'bg-rose-saumon-bg'; return (<div className={`flex justify-between items-center p-3 rounded-xl ${genderBg}`}><p className="font-semibold text-text-primary">{name.prénom}{lastName && <span className="ml-2 font-normal text-text-secondary">{lastName}</span>}</p><button onClick={() => onToggleLike(name.id)}><HeartIcon className={`w-7 h-7 transition-all duration-200 transform ${isLiked ? 'text-rose-saumon scale-100' : 'text-slate-300 dark:text-slate-500 scale-90 hover:text-rose-saumon/70 hover:scale-110'}`} /></button></div>); };
-const NameDetailScreen = ({ name, profile, updateProfile, onBack }) => {
-    const isLiked = profile.likedNames.includes(name.id);
-    const handleLike = () => {
-        const currentLiked = new Set(profile.likedNames);
-        if (currentLiked.has(name.id)) {
-            currentLiked.delete(name.id);
-        } else {
-            currentLiked.add(name.id);
-        }
-        updateProfile({ likedNames: Array.from(currentLiked) });
-    };
-    const genderBg = name.genre === 'M' ? 'bg-bleu-layette-bg' : 'bg-rose-saumon-bg';
-    return (
-        <AnimatedPage>
-            <button onClick={onBack} className="text-lavande-poudre font-semibold flex items-center space-x-2 group mb-4">
-                <span className="transform transition-transform group-hover:-translate-x-1">&larr;</span>
-                <span>Retour</span>
-            </button>
-            <Card className={`text-center ${genderBg}`}>
-                <h1 className="text-6xl font-bold text-text-primary">
-                    {name.prénom}
-                    {profile.lastName && <span className="text-5xl ml-3 text-text-secondary">{profile.lastName}</span>}
-                </h1>
-                <p className="italic text-text-secondary my-4 text-lg">"{name.signification}"</p>
-                <p className="text-text-secondary">Origine(s): {name.origines.join(', ')}</p>
-                <div className="flex justify-center items-center mt-8 pt-6 border-t border-subtle-border">
-                    <button
-                        onClick={handleLike}
-                        aria-label={isLiked ? 'Retirer des favoris' : 'Ajouter aux favoris'}
-                    >
-                        <HeartIcon
-                            isLiked={isLiked}
-                            className={`h-16 w-16 transition-all transform hover:scale-110 active:scale-100 ${isLiked ? 'text-rose-saumon' : 'text-slate-300 dark:text-slate-600'}`}
-                        />
-                    </button>
-                </div>
-            </Card>
-        </AnimatedPage>
-    );
-};
 const ThemeToggle = ({ theme, setTheme }) => { const toggleTheme = () => setTheme(theme === 'light' ? 'dark' : 'light'); return (<div className="flex items-center justify-between"><span className="text-text-primary font-semibold">Mode Sombre</span><button onClick={toggleTheme} className={`relative inline-flex h-6 w-11 flex-shrink-0 cursor-pointer rounded-full border-2 border-transparent transition-colors duration-200 ease-in-out focus:outline-none focus:ring-2 focus:ring-lavande-poudre focus:ring-offset-2 dark:focus:ring-offset-slate-800 ${theme === 'dark' ? 'bg-lavande-poudre' : 'bg-gray-200'}`}><span aria-hidden="true" className={`pointer-events-none inline-block h-5 w-5 transform rounded-full bg-white shadow ring-0 transition duration-200 ease-in-out ${theme === 'dark' ? 'translate-x-5' : 'translate-x-0'}`}><span className={`absolute inset-0 flex h-full w-full items-center justify-center transition-opacity ${theme === 'dark' ? 'opacity-0 duration-100 ease-out' : 'opacity-100 duration-200 ease-in'}`}><SunIcon className="h-3 w-3 text-gray-400" /></span><span className={`absolute inset-0 flex h-full w-full items-center justify-center transition-opacity ${theme === 'dark' ? 'opacity-100 duration-200 ease-in' : 'opacity-0 duration-100 ease-out'}`}><MoonIcon className="h-3 w-3 text-lavande-poudre" /></span></span></button></div>); };
 
 
@@ -1238,6 +674,7 @@ const StyleInjector = () => {
     }, []);
     return null;
 }
+export { AnimatedPage, Card, EyeIcon, HeartIcon, TrophyIcon, UndoIcon, FilterIcon, SwipeableCard, FavoriteItem, Button, LoadingSpinner, ThemeToggle, LogoutIcon, UserPlusIcon, Modal, Chip, FAB, SearchIcon, RangeSlider, BackgroundDecorations, SunIcon, MoonIcon, TrashIcon };
 
 // --- FINAL APP EXPORT ---
 const EnhancedApp = () => (
