+++ src/screens/DashboardScreen.js
@@ -1,0 +1,5 @@
+import { useMemo } from 'react';
+import { AnimatedPage, Card, EyeIcon, HeartIcon, TrophyIcon } from '../App';
+const DashboardScreen = ({ profile, allNames }) => { const liked = useMemo(() => profile.likedNames.map(id => allNames.find(n => n.id === id)).filter(Boolean), [profile.likedNames, allNames]); const m = useMemo(() => liked.filter(n => n.genre === 'M').length, [liked]); const boyPercent = liked.length > 0 ? Math.round((m / liked.length) * 100) : 0; const medalColors = ['#FFD700', '#C4C4C4', '#D9A77A']; return (<AnimatedPage className="space-y-6"><div className="text-center"><h1 className="text-3xl font-bold text-text-primary">Bonjour, {profile.displayName} !</h1><p className="text-text-secondary mt-1">Voici le résumé de vos découvertes.</p></div><Card><h2 className="font-bold text-lg mb-4 text-center text-text-primary">Statistiques</h2><div className="flex justify-around text-center"><div className="flex flex-col items-center space-y-1"><EyeIcon className="h-8 w-8 text-lavande-poudre" /><p className="text-3xl font-bold text-text-primary">{profile.seenNames.length}</p><p className="text-text-secondary text-sm">Prénoms vus</p></div><div className="flex flex-col items-center space-y-1"><HeartIcon className="h-8 w-8 text-rose-saumon" /><p className="text-3xl font-bold text-text-primary">{profile.likedNames.length}</p><p className="text-text-secondary text-sm">Coups de cœur</p></div></div></Card><Card><h2 className="font-bold text-lg mb-3 text-center text-text-primary">Répartition Garçon/Fille</h2><div className="w-full bg-rose-saumon-bg rounded-full h-4 relative overflow-hidden"><div className="bg-bleu-layette-bg h-4 rounded-l-full" style={{ width: `${boyPercent}%` }}></div></div><div className="flex justify-between text-xs text-text-secondary mt-1"><span>Garçon {boyPercent}%</span><span>Fille {100 - boyPercent}%</span></div></Card><Card><h2 className="font-bold text-lg mb-3 text-center text-text-primary">Nos petits trésors 🏆</h2>{profile.likedNames.length > 0 ? (<ul className="space-y-2">{profile.likedNames.slice(0, 3).map((nameId, index) => { const name = allNames.find(n => n.id === nameId); if (!name) return null; const genderBg = name.genre === 'M' ? 'bg-bleu-layette-bg' : 'bg-rose-saumon-bg'; return (<li key={name.id} className={`flex items-center space-x-4 p-3 rounded-xl ${genderBg} border-l-4`} style={{ borderColor: medalColors[index] }}><TrophyIcon color={medalColors[index]} className="h-7 w-7 flex-shrink-0" /><p className="font-semibold text-text-primary"><span className="text-lg">{name.prénom}</span>{profile.lastName && <span className="text-base ml-2 text-text-secondary">{profile.lastName}</span>}</p></li>) })}</ul>) : (<p className="text-center text-text-secondary py-4">Aimez des prénoms pour voir vos trésors !</p>)}</Card></AnimatedPage>); };
+
+export default DashboardScreen;
