+++ src/screens/FavoritesListScreen.js
@@ -1,0 +1,5 @@
+import { useState, useEffect } from 'react';
+import { AnimatedPage, Card, HeartIcon, FavoriteItem } from '../App';
+const FavoritesListScreen = ({ profile, allNames, updateProfile, setDetailedName }) => { const [favoriteItems, setFavoriteItems] = useState([]); const [draggedItem, setDraggedItem] = useState(null); useEffect(() => { setFavoriteItems([...profile.likedNames].reverse()); }, [profile.likedNames]); const handleRemove = (nameId) => { const newLikedNames = profile.likedNames.filter(id => id !== nameId); updateProfile({ likedNames: newLikedNames }); }; const handleDragStart = (e, index) => { setDraggedItem(index); e.dataTransfer.effectAllowed = 'move'; }; const handleDragOver = (e, index) => { e.preventDefault(); }; const handleDrop = (e, dropIndex) => { const draggedId = favoriteItems[draggedItem]; const newItems = [...favoriteItems]; newItems.splice(draggedItem, 1); newItems.splice(dropIndex, 0, draggedId); setFavoriteItems(newItems); updateProfile({ likedNames: [...newItems].reverse() }); }; const handleDragEnd = () => { setDraggedItem(null); }; return (<AnimatedPage><h1 className="text-3xl font-bold text-text-primary mb-2">Vos Trésors</h1><p className="text-text-secondary mb-6 -mt-1">Organisez votre liste en glissant-déposant les prénoms.</p>{favoriteItems.length > 0 ? (<ul className="space-y-3">{favoriteItems.map((id, index) => { const name = allNames.find(n => n.id === id); if (!name) return null; return (<FavoriteItem key={id} name={name} lastName={profile.lastName} index={index} isBeingDragged={draggedItem === index} onRemove={handleRemove} onViewDetails={setDetailedName} onDragStart={(e) => handleDragStart(e, index)} onDragOver={(e) => handleDragOver(e, index)} onDrop={(e) => handleDrop(e, index)} onDragEnd={handleDragEnd} />); })}</ul>) : (<Card className="text-center"><HeartIcon className="h-16 w-16 mx-auto text-slate-300 dark:text-slate-600" /><h3 className="text-xl font-semibold text-text-primary mt-4">Aucun trésor pour l'instant</h3><p className="text-text-secondary mt-2">Allez dans "Trouvailles" pour découvrir des prénoms !</p></Card>)}</AnimatedPage>); };
+
+export default FavoritesListScreen;
