+++ src/screens/FilterScreen.js
@@ -1,0 +1,5 @@
+import { useMemo } from 'react';
+import { AnimatedPage, Card, Chip, RangeSlider, FAB, SearchIcon } from '../App';
+const FilterScreen = ({ filter: current, origins, onSetFilter, onLaunch, allNames }) => { const handleOriginToggle = (origin) => { const newOrigins = [...current.origins]; const index = newOrigins.indexOf(origin); if (index > -1) { newOrigins.splice(index, 1); } else { newOrigins.push(origin); } onSetFilter({ ...current, origins: newOrigins }); }; const resetOrigins = () => { onSetFilter({ ...current, origins: [] }); }; const handleLengthTypeChange = (type) => { onSetFilter({ ...current, length: { ...current.length, type } }); }; const handleLengthRangeChange = ({ min, max }) => { onSetFilter({ ...current, length: { ...current.length, type: 'Personnalisé', min, max } }); }; const lengthTypes = [{ key: 'Tous', label: 'Toutes' }, { key: 'Court', label: 'Courts (3-4)' }, { key: 'Moyen', label: 'Moyens (5-6)' }, { key: 'Long', label: 'Longs (7+)' }, { key: 'Personnalisé', label: 'Personnalisé' },]; const filteredCount = useMemo(() => { if (!allNames) return 0; return allNames.filter(n => { const genderMatch = (current.gender === 'Tous' || n.genre === current.gender); const originMatch = (current.origins.length === 0 || (n.filtre_global && current.origins.includes(n.filtre_global))); const nameLength = n.prénom.length; const { type, min, max } = current.length; let lengthMatch = true; switch (type) { case 'Court': lengthMatch = nameLength >= 3 && nameLength <= 4; break; case 'Moyen': lengthMatch = nameLength >= 5 && nameLength <= 6; break; case 'Long': lengthMatch = nameLength >= 7; break; case 'Personnalisé': lengthMatch = nameLength >= min && nameLength <= max; break; default: lengthMatch = true; } return genderMatch && originMatch && lengthMatch; }).length; }, [allNames, current]); const totalCount = allNames.length; return (<> <AnimatedPage className="space-y-6 pb-24"> <Card> <h2 className="text-xl font-bold mb-4 text-text-primary">Genre</h2> <div className="flex justify-center space-x-2"> <Chip label="Garçon" selected={current.gender === 'M'} onClick={() => onSetFilter({ ...current, gender: 'M' })} /> <Chip label="Fille" selected={current.gender === 'F'} onClick={() => onSetFilter({ ...current, gender: 'F' })} /> <Chip label="Tous" selected={current.gender === 'Tous'} onClick={() => onSetFilter({ ...current, gender: 'Tous' })} /> </div> </Card> <Card> <div className="flex justify-between items-center mb-4"> <h2 className="text-xl font-bold text-text-primary">Origine(s) {current.origins.length > 0 ? `(${current.origins.length})` : ''}</h2> {current.origins.length > 0 && (<button onClick={resetOrigins} className="text-sm font-semibold text-lavande-poudre hover:underline">Tout désélectionner</button>)} </div> <div className="max-h-48 overflow-y-auto pr-2"> <div className="flex flex-wrap gap-2"> <Chip label="Toutes" selected={current.origins.length === 0} onClick={resetOrigins} /> {origins.map(origin => (<Chip key={origin} label={origin} selected={current.origins.includes(origin)} onClick={() => handleOriginToggle(origin)} />))} </div> </div> </Card> <Card> <h2 className="text-xl font-bold mb-4 text-text-primary">Longueur du prénom</h2> <div className="flex flex-wrap justify-center gap-2"> {lengthTypes.map(({ key, label }) => (<Chip key={key} label={label} selected={current.length.type === key} onClick={() => handleLengthTypeChange(key)} />))} </div> {current.length.type === 'Personnalisé' && (<div className="mt-6 animate-fade-in-up"> <RangeSlider min={2} max={15} currentMin={current.length.min} currentMax={current.length.max} onChange={handleLengthRangeChange} /> </div>)} </Card> <Card className="text-center mt-6 bg-card-bg/80 backdrop-blur-sm"> <p className="text-text-secondary text-sm sm:text-base"> <span className="font-bold text-text-primary">{filteredCount}</span> {filteredCount > 1 ? "prénoms correspondent" : "prénom correspond"} <br /> sur <span className="font-bold text-text-primary">{totalCount}</span> prénoms disponibles. </p> </Card> </AnimatedPage> <FAB icon={SearchIcon} onClick={onLaunch} ariaLabel="Appliquer les filtres" /> </>); };
+
+export default FilterScreen;
